### 2025-10-24 #9300 @editorialbot
[PRE REVIEW]: Magrathea v2: A planetary interior modeling platform in C++
**Submitting author:** @DavidRRice (<a href="http://orcid.org/0000-0001-6009-8685">David Rice</a>)
**Repository:** https://github.com/Huang-CL/Magrathea
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0.0
