### 2025-10-24 #9300 @warrickball
Hi @DavidRRice, and thanks for your submission! Before I add this to your submission queue, could you fix the incorrect ORCID in the paper? I presume Jason Steffen's should be 0000-0003-2202-384**7**.

The paper is also a touch on the long side. [Our docs advise](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) that papers should be between 250–1000 words. That's not a hard limit but it would be good if you could remove a few hundred. At a glance at the `.md` source, the text is a bit verbose and repetitive, e.g. in the summary you have

> The result is a more versatile platform that supports a winder [sic] range of planetary compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.

and then at the end of the updates

> By expanding the physics library, adding composition retrieval solvers, and improving usability, the code now enables a wider range of applications than in its initial release.

That's not the only example, though.

Regarding the codebase itself, to evaluate whether this meets our requirements of [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), we'd like to inspect the code that has been added since the 2022 publication. I've been looking at the `git diff` between the top of the default branch and commit https://github.com/Huang-CL/Magrathea/commit/c18ec2bfab328200822caf7b1c8f548801c3051b but if there are any specific blocks of code, PRs, or anything else to guide us, that would be useful. It appears that the most substantial additions are the composition finder and additional of more EoSes.
