### 2025-10-25 6d64c70
Update paper.md
