+++ docs/paper.md
@@ -27,7 +27,7 @@ authors:
     orcid: 0000-0002-0957-3177
     affiliation: 6
   - name: Jason H. Steffen
-    orcid: 0000-0003-2202-384
+    orcid: 0000-0003-2202-3847
     affiliation: 3
   - name: Allona Vazan
     orcid: 0000-0001-9504-3174
@@ -52,18 +52,18 @@ bibliography: paper.bib
 
 # Summary
 
-Magrathea is an open-source C++ code for modeling the internal structure of differentiated planets. The initial release, @Huang:2022, introduced the base solver, a modular framework for defining equations of state (EOSs) used within phase diagrams for each differentiated layer, and outlined a series of planned extensions. Many of those updates are now implemented here, alongside new physics, solvers, and usability improvements. The result is a more versatile platform that supports a winder range of planetary compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.
+Magrathea is an open-source C++ code for modeling the internal structure of differentiated planets. The initial release, @Huang:2022, introduced the base solver, a modular framework for defining equations of state (EOSs) used within phase diagrams for each differentiated layer, and outlined a series of planned extensions. Many of those updates are now implemented. Magrathea v2 is a more versatile platform that supports a winder range of compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.
 
 
 # Statement of need
 
-Constraining a planet’s composition is essential for understanding its formation and evolution. Observations of mass and radius alone are not sufficient, since many different interiors can yield the same bulk density. Interior structure solvers are therefore essential tools for exploring this degeneracy and constraining the possible distributions of iron, silicate, volatiles, and atmosphere. With observational programs routinely measuring the densities of small to large planets, researchers require codes with models that are transparent and flexible—able to adapt to our changing understanding of planet mineralogy. 
+Constraining a planet’s composition is essential for understanding its formation and evolution. Observations of mass and radius alone are not sufficient, since many different interiors can yield the same bulk density. Interior structure solvers are therefore essential tools for constraining the possible compositions. With observational programs routinely measuring the densities of small to large planets, researchers require codes with models that are transparent and flexible—able to adapt to our changing understanding of planet mineralogy. 
 
 Magrathea is designed as such a platform. Rather than enforcing a fixed planet model, Magrathea provides a framework in which users can define their own phase diagrams, equations of state (EOSs), and thermal profiles. This adaptability has led to broad uptake: Magrathea has been used to generate mass–radius diagrams and infer interiors of observed planets [@Desai:2024; @Daspute:2025; @Rice:2025; @Taylor:2025], to connect theoretical composition models to observables [@Childs:2023; @Dou:2024; @Steffen:2025], and to incorporate new high-pressure equation of state (EOS) measurements into planetary modeling [@Huang:2021]. A list of other open source interior models can be found in @Acuna:2025. With the continued expansion of the physics and usability, Magrathea helps the community keep up with the growing precision of exoplanet observations and experimental constraints on planetary materials.
 
 # Summary of the base code
 
[…]
