### 2025-10-26 #9300 @DavidRRice
Thanks @warrickball. I should have mentioned more clearly v2.0.0-beta was from April, the better comparison is to the current paper branch included below.  We decrease the size of some of the large tables we included for portability, so the text is cut, but much more code was added. I'm glad such a great opportunity is oversubscribed, but a bit surprised by the 6 months. We were definitely hoping to get this out to our users quickly as several works are expected this next year using the updated code. Is it acceptable to post it on the arXiv and any guidelines from JOSS for doing so?

```
cloc --git-diff-all v1.0.0 paper
      32 text files.
      80 text files.
      55 files ignored.

Language                     files          blank        comment           code
C++
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           8           1121            612           5887
 removed                         6            810            374           4194
Markdown
 same                            0              0              0              0
 modified                        0              0              0              0
 added                          12            403              0           1054
 removed                         5            106              0            199
C/C++ Header
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           8             99            175            370
 removed                         6             68            105            261
TeX
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           1             38              0            550
 removed                         0              0              0              0
Python
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           5             61             55            159
 removed                         2             31             33             82
YAML
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           2              4              4             32
 removed                         0              0              0              0
make
 same                            0              0              0              0
 modified                        0              0              0              0
 added                           1              6              1             16
 removed                         1              6              1             16
SUM:
 same                            0              0              0              0
 modified                        0              0              0              0
 added                          37           1732            847           8068
 removed                        20           1021            513           4752
```
