### 2025-10-27 f24d905
Update paper.md
