+++ README.md
@@ -6,7 +6,7 @@
 
 Planet interior structure code for astronomers, planetary scientists, mice, and more. 
 
-[➡️ Read the Docs](https://magrathea.readthedocs.io/en/latest/) [➡️ v1 Pub.](https://arxiv.org/abs/2201.03094) [➡️ v2 Pub.](https://arxiv.org/abs/2510.23733)
+[➡️ Read the Docs](https://magrathea.readthedocs.io/en/latest/)&nbsp;&nbsp;&nbsp;&nbsp;[➡️ v1 Pub.](https://arxiv.org/abs/2201.03094)&nbsp;&nbsp;&nbsp;&nbsp;[➡️ v2 Pub.](https://arxiv.org/abs/2510.23733)
 
 ## What is this repository for? ##
 
