### 2025-10-29 31ac6d0
Update paper.md
