### 2025-10-30 714d66f
Update paper.md
