+++ docs/paper.md
@@ -37,13 +37,13 @@ affiliations:
    index: 1
  - name: Shanghai Astronomical Observatory, Chinese Academy of Sciences, Shanghai 200030, People’s Republic of China
    index: 2
- - name: Department of Physics and Astronomy, University of Nevada, Las Vegas, 4505 South Maryland Parkway, Las Vegas, NV 89154, US
+ - name: Department of Physics and Astronomy, University of Nevada, Las Vegas, 4505 South Maryland Parkway, Las Vegas, NV 89154, USA
    index: 3
  - name: Department of Physics, Ariel University, Ariel 40700, Israel
    index: 4
- - name: Department of Astronomy, University of Wisconsin–Madison, 475 N. Charter Street, Madison, WI, 53706, USA
+ - name: Department of Astronomy, University of Wisconsin–Madison, 475 N. Charter Street, Madison, WI 53706, USA
    index: 5
- - name: Department of Earth and Space Science, University of Washington, Seattle, WA 98195, United States
+ - name: Department of Earth and Space Science, University of Washington, Seattle, WA 98195, USA
    index: 6 
 
 date: 23 October 2025
@@ -63,13 +63,13 @@ Magrathea is designed as such a platform. Rather than enforcing a fixed planet m
 
 # Summary of the base code
 
-The core solver of Magrathea is a one-dimensional, spherically symmetric integrator of the equations of hydrostatic equilibrium, mass continuity, and energy transport written in C++. For a user-defined planet consisting of up to four differentiated layers, the code integrates inward and outward solutions using a shooting-to-fitting-point method with adaptive Runge–Kutta–Fehlberg stepping. The solver returns the radius of the planet, the radii of each compositional boundary, and profiles of pressure, temperature, density, and phase as functions of enclosed mass. Solving one planet taskes approximately one second for most configurations.
+The core solver of Magrathea is a one-dimensional, spherically symmetric integrator of the equations of hydrostatic equilibrium, mass continuity, and energy transport written in C++. For a user-defined planet consisting of up to four differentiated layers, the code integrates inward and outward solutions using a shooting-to-fitting-point method with adaptive Runge–Kutta–Fehlberg stepping. The solver returns the radius of the planet, the radii of each compositional boundary, and profiles of pressure, temperature, density, and phase as functions of enclosed mass. Solving one planet takes approximately one second for most configurations.
 
 A key design choice is **modularity**:
 
 - A large variety of EOS forms are supported in `EOS.cpp`, including Birch–Murnaghan, Vinet, Holzapfel, Keane, van der Waals gases, and tabulated.
 - Parameters for each material's EOS are defined and stored in a library of 70+ EOSs in `EOSlist.cpp`.
-- Phase diagrams for each layer define which material is used at a given P-T condition in `phase.cpp`.
+- Phase diagrams for each layer define which material is used at a given P–T condition in `phase.cpp`.
 
 Magrathea offers **nine run modes** through human-readable `.cfg` files:  
 1. **Full solver** takes masses for each layer and returns the planet’s radius and interior profiles.  
@@ -82,7 +82,7 @@ Magrathea offers **nine run modes** through human-readable `.cfg` files:
 8. **Iterated EOS modification** with full solver.  
 9. **MCMC composition retrieval** for probabilistic inference given mass, radius, and corresponding uncertainties.  
 
-This modularity and range of modes makes Magrathea not just a solver but a platform for exploring interior models.
+This modularity and range of modes make Magrathea not just a solver but a platform for exploring interior models.
 
 
[…]
