### 2025-11-04 c4a34f5
Documentation: JOSS preprint & ReadTheDocs (#29)
To detail the many expansions of the code, we've included:
Preprint submitted to JOSS: https://arxiv.org/abs/2510.23733
Read The Docs: https://magrathea.readthedocs.io/en/latest/
