+++ docs/paper.bib
@@ -1,0 +1,588 @@
+@ARTICLE{Huang:2022,
+       author = {{Huang}, Chenliang and {Rice}, David R. and {Steffen}, Jason H.},
+        title = "{MAGRATHEA: an open-source spherical symmetric planet interior structure code}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {equation of state, software: public release, planets and satellites: composition, planets and satellites: general, planets and satellites: interiors, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = jul,
+       volume = {513},
+       number = {4},
+        pages = {5256-5269},
+          doi = {10.1093/mnras/stac1133},
+archivePrefix = {arXiv},
+       eprint = {2201.03094},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.513.5256H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Taylor:2025,
+       author = {{Taylor}, Jake and {Radica}, Michael and {Chatterjee}, Richard D. and {Hammond}, Mark and {Meier}, Tobias and {Aigrain}, Suzanne and {MacDonald}, Ryan J. and {Albert}, Loic and {Benneke}, Bj{\"o}rn and {Coulombe}, Louis-Philippe and {Cowan}, Nicolas B. and {Dang}, Lisa and {Doyon}, Ren{\'e} and {Flagg}, Laura and {Johnstone}, Doug and {Kaltenegger}, Lisa and {Lafreni{\`e}re}, David and {Pelletier}, Stefan and {Piaulet-Ghorayeb}, Caroline and {Rowe}, Jason F. and {Roy}, Pierre-Alexis},
+        title = "{JWST NIRISS transmission spectroscopy of the super-Earth GJ 357b, a favourable target for atmospheric retention}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {planets and satellites: atmospheres, planets and satellites: individual: GJ 357 b, planets and satellites: terrestrial planets, Earth and Planetary Astrophysics, Solar and Stellar Astrophysics},
+         year = 2025,
+        month = jul,
+       volume = {540},
+       number = {4},
+        pages = {3677-3692},
+          doi = {10.1093/mnras/staf894},
+archivePrefix = {arXiv},
+       eprint = {2505.24462},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025MNRAS.540.3677T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Rice:2025,
+       author = {{Rice}, David R. and {Huang}, Chenliang and {Steffen}, Jason H. and {Vazan}, Allona},
+        title = "{Uncertainties in the Inference of Internal Structure: The Case of TRAPPIST-1 f}",
+      journal = {The Astrophysical Journal},
+     keywords = {Exoplanets, Exoplanet structure, Planetary interior, Hydrosphere, Extrasolar rocky planets, 498, 495, 1248, 770, 511, Earth and Planetary Astrophysics},
+         year = 2025,
+        month = jun,
+       volume = {986},
+       number = {1},
+          eid = {2},
+        pages = {2},
+          doi = {10.3847/1538-4357/add34b},
+archivePrefix = {arXiv},
+       eprint = {2504.16201},
+ primaryClass = {astro-ph.EP},
[…]
