+++ docs/paper.md
@@ -1,0 +1,126 @@
+---
+title: 'Magrathea v2: A planetary interior modeling platform in C++'
+tags:
+  - C++
+  - astronomy
+  - planets
+  - exoplanets
+  - interiors
+authors:
+  - name: David R. Rice
+    orcid: 0000-0001-6009-8685
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: 1
+  - name: Chenliang Huang
+    orcid: 0000-0001-9446-6853
+    affiliation: 2
+  - name: Robert Royer
+    orcid: 0000-0002-0900-0192
+    affiliation: 3
+  - name: Mangesh Daspute
+    orcid: 
+    affiliation: 4
+  - name: Krishang Mittal
+    orcid: 
+    affiliation: 5
+  - name: Baptiste Journaux
+    orcid: 0000-0002-0957-3177
+    affiliation: 6
+  - name: Jason H. Steffen
+    orcid: 0000-0003-2202-3847
+    affiliation: 3
+  - name: Allona Vazan
+    orcid: 0000-0001-9504-3174
+    affiliation: 1
+affiliations:
+ - name: Astrophysics Research Center (ARCO), Department of Natural Sciences, The Open University of Israel, Raanana 4353701, Israel
+   index: 1
+ - name: Shanghai Astronomical Observatory, Chinese Academy of Sciences, Shanghai 200030, People’s Republic of China
+   index: 2
+ - name: Department of Physics and Astronomy, University of Nevada, Las Vegas, 4505 South Maryland Parkway, Las Vegas, NV 89154, USA
+   index: 3
+ - name: Department of Physics, Ariel University, Ariel 40700, Israel
+   index: 4
+ - name: Department of Astronomy, University of Wisconsin–Madison, 475 N. Charter Street, Madison, WI 53706, USA
+   index: 5
+ - name: Department of Earth and Space Science, University of Washington, Seattle, WA 98195, USA
+   index: 6 
+
+date: 23 October 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Magrathea is an open-source C++ code for modeling the internal structure of differentiated planets. The initial release, @Huang:2022, introduced the base solver, a modular framework for defining equations of state (EOSs) used within phase diagrams for each differentiated layer, and outlined a series of planned extensions. Many of those updates are now implemented. Magrathea v2 is a more versatile platform that supports a wider range of compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.
+
+
+# Statement of need
+
+Constraining a planet’s composition is essential for understanding its formation and evolution. Observations of mass and radius alone are not sufficient, since many different interiors can yield the same bulk density. Interior structure solvers are therefore essential tools for constraining the possible compositions. With observational programs routinely measuring the densities of small to large planets, researchers require codes with models that are transparent and flexible—able to adapt to our changing understanding of planet mineralogy. 
+
+Magrathea is designed as such a platform. Rather than enforcing a fixed planet model, Magrathea provides a framework in which users can define their own phase diagrams, equations of state (EOSs), and thermal profiles. This adaptability has led to broad uptake: Magrathea has been used to generate mass–radius diagrams and infer interiors of observed planets [@Desai:2024; @Daspute:2025; @Rice:2025; @Taylor:2025], to connect theoretical composition models to observables [@Childs:2023; @Dou:2024; @Steffen:2025], and to incorporate new high-pressure equation of state (EOS) measurements into planetary modeling [@Huang:2021]. A list of other open source interior models can be found in @Acuna:2025. With the continued expansion of the physics and usability, Magrathea helps the community keep up with the growing precision of exoplanet observations and experimental constraints on planetary materials.
+
[…]
