+++ docs/rtd/index.md
@@ -1,0 +1,22 @@
+# Magrathea Documentation
+
+Welcome! This site explains how to **install**, **run**, and **extend** Magrathea.  
+It also hosts an auto-generated **C++ API Reference** from the source code.
+
+```{toctree}
+:maxdepth: 2
+:caption: User Guide
+
+getting-started
+user-guide/run-modes
+models/phasediagrams
+models/eos
+```
+
+```{toctree}
+:maxdepth: 2
+:caption: Reference
+
+reference/api-reference-guide
+docs_api/library_root
+```

[…]
