### 2025-11-08 b8fcc4f
Updated EoSList.cpp
