### 2025-11-09 33cf621
Update Graphite EOS
