+++ src/EOSlist.h
@@ -3,7 +3,7 @@
 
 #include "EOS.h"
 
-extern EOS *Fe_liquid, *Fe_liquid2, *Fe_fcc, *Fe_bcc, *Fe_hcp, *Fe_hcp2, *Fe_Seager, *Fe_hcp3, *Fe_7Si, *Fe_15Si, *Fe_Dummy, *Si_Pv_Shim, *Si_Pv, *Si_PPv, *Si_PPv_Sakai, *Si_PREM, *Si_BM2fit, *Si_Seager, *Si_Dummy, *Si_liquid, *Si_Liquid_Wolf, *Fo, *Wds, *Rwd, *Akm, *Pv_Doro, *PPv_Doro, *Fo_Sotin, *En, *Mw, *Ice_Seager, *H2O_AQUA, *H2O_SeaFreeze, *Water_SF, *Water_Brown, *Water_IAPWS, *Water_ExoPlex, *Water, *Water_sc_Mazevet, *IceIh_SF, *IceII_SF, *IceIII_SF, *IceV_SF, *IceVI_SF, *IceIh, *IceIh_ExoPlex, *IceVI_ExoPlex, *IceVI_Bezacier, *IceVII_ExoPlex, *IceVII_Bezacier, *IceVII, *IceVIIp, *IceVII_FFH2004, *IceVII_FFH2004fit, *IceVII_FFH2004BM, *IceVII_Fei, *IceVII_FFH2004T, *IceX_HS, *IceX, *IceZeng2013FFH, *IceZeng2013FMNR, *Ice_Dummy, *Gas, *Gas_iso, *Gas_hhe, *watervapor, *Water_Vap_IAPWS, *vdW_H2, *vdW_He, *vdW_H2O, *vdW_CH4, *vdW_NH3, *vdW_CO2, *Gold, *Plat, *Graph, *Diam, *BC8, *SiC_B3_Vinet, *SiC_B1_Vinet, *Anorthite, *Spinel, *Fayalite, *Fe_Wadsleyite, *Fe_Ringwoodite, *Enstatite, *Ferrosilite, *Diopside, *Hedenbergite, *HP_clinopyroxene, *Mg_Akimotoite, *Fe_Akimotoite, *Pyrope, *Mg_Majorite, *Almandine, *Grossular, *Quartz, *Coesite, *Stishovite, *Seifertite, *Fe_Post_Perovskite, *Fe_Perovskite, *HP_Clinoferrosilite, *Periclase, *Wustite, *Kyanite, *Nepheline, *Graph_Lowitzer;
[…]
