### 2025-11-12 06c0884
Add temperature dependent adiabatic index
