### 2025-11-13 e9b4444
minor change to fallback of gamma_shomate
