### 2025-11-18 ffc667f
Update EOSlist.cpp
Edit to V0 for Graphite_Lowitzer to reflect four atoms per unit cell.
