### 2025-11-20 4734681
Update EOSlist.cpp
