### 2025-11-20 a6575f2
Update paper.md
