+++ docs/paper.bib
@@ -585,4 +585,19 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Lowitzer:2006,
+       author = {{Lowitzer}, S. and {Winkler}, B. and {Tucker}, M.},
+        title = "{Thermoelastic behavior of graphite from in situ high-pressure high-temperature neutron diffraction}",
+      journal = {Physical Review B},
+     keywords = {65.40.-b, Thermal properties of crystalline solids},
+         year = 2006,
+        month = jun,
+       volume = {73},
+       number = {21},
+          eid = {214115},
+        pages = {214115},
+          doi = {10.1103/PhysRevB.73.214115},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2006PhRvB..73u4115L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
