### 2025-11-20 4fcccee
Update paper.md
