### 2026-01-28 #9300 @DavidRRice
@warrickball Hi, I was hoping for a quick update on the timeline of receiving an editor.
