### 2026-02-01 #9300 @warrickball
Hi again @DavidRRice, I apologise for the ongoing wait but no, an editor has not become available yet and is unlikely to for at least 3 months, and suspect more. Sorry!

In the meantime, we recently [announced new changes to our requirements](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future), which now appear in our docs on [what the paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html), and [scope definition](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance), and we've decided to re-evaluate existing submissions in pre-review against the new criteria.

Could you update the paper to match the new structure? This may be by re-arranging some of the existing material or writing new material. The maximum word count has increased from 1000 words to 1750 to accommodate the extra requirements. Your submission will keep its place in the submission queue. Just post here (or regenerate the PDF) when you've made these changes, and I'll make any comments necessary at this point.

We want to be transparent that under these new criteria, some submissions that were previously in scope may now be deemed out of scope, and we may choose not to proceed with review.  The editorial board discussed this issue for some time and our consensus was that this choice will give JOSS's volunteer editors and reviewers the most clarity about what requirements to apply and support our commitment to ensuring JOSS publications recognise genuine scholarly contributions.
