### 2026-02-01 #9300 @warrickball
> I believe we still meet the scope and significance.

I'm going to go ahead and say that yes, I also believe this should meet the new requirements. There's clear evidence of collaborative development, the code changes are sufficiently substantial and I think the research impact is demonstrated but the citations to the original MAGRATHEA paper.

I recommend that when you prepare the "Research Impact" section, try to highlight publications (if any) that have used features present in v2 and not just v1, just to motivate that the code changes have themselves had impact.
