+++ src/EOSlist.cpp
@@ -279,7 +279,7 @@ EOS *Water = new EOS("Water (Valencia)", Water_array, sizeof(Water_array)/2/size
 // -----------------------------------
 // Water & Super Critical, Brown 2018, Fluid Phase Equilib., 463, 18, tabulated by SeaFreeze
 // DEFAULT
-EOS *Water_Brown = new EOS("Water SC (Brown)", "./tabulated/water_Brown.txt");
+EOS *Water_Brown = new EOS("Water Liq&SC (Brown)", "./tabulated/water_Brown.txt");
 
 // -----------------------------------
 // Supercritical water. Mazevet et al. 2019, A&A 621
@@ -1384,6 +1384,7 @@ void FINVER(double F, int N, double &X, double &XDF, double &XDFF)
     XDF  = t1 * RT;
     XDFF = t2 * RT + pow(t1,2) * (R2 - 2.0*RT) / t;
 }
+
 
 
 
