### 2026-04-28 528d4f8
Update JOSS Paper
Updated JOSS Paper following new required sections
