+++ docs/paper.bib
@@ -585,4 +585,112 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Lowitzer:2006,
+       author = {{Lowitzer}, S. and {Winkler}, B. and {Tucker}, M.},
+        title = "{Thermoelastic behavior of graphite from in situ high-pressure high-temperature neutron diffraction}",
+      journal = {Physical Review B},
+     keywords = {65.40.-b, Thermal properties of crystalline solids},
+         year = 2006,
+        month = jun,
+       volume = {73},
+       number = {21},
+          eid = {214115},
+        pages = {214115},
+          doi = {10.1103/PhysRevB.73.214115},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2006PhRvB..73u4115L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Baumeister:2025,
+       author = {{Baumeister}, Philipp and {Miozzi}, Francesca and {Guimond}, Claire Marie and {Steinmeyer}, Marie-Luise and {Dorn}, Caroline and {Karato}, Shun-Ichiro and {Bolmont}, {\'E}meline and {Revol}, Alexandre and {Thamm}, Alexander and {Noack}, Lena},
+        title = "{Fundamentals of Interior Modelling and Challenges in the Interpretation of Observed Rocky Exoplanets}",
+      journal = {Space Science Reviews},
+     keywords = {Earth and Planetary Astrophysics, Geophysics},
+         year = 2025,
+        month = dec,
+       volume = {221},
+       number = {8},
+          eid = {123},
+        pages = {123},
+          doi = {10.1007/s11214-025-01248-5},
+archivePrefix = {arXiv},
+       eprint = {2511.10269},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025SSRv..221..123B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Kroft:2026,
+       author = {{Kroft}, Maxwell A. and {Beatty}, Thomas G. and {Salzer}, Joseph M. and {Zwicker}, Claire and {Triantafillides}, Anastasia and {Becker}, Juliette and {Soares-Furtado}, Melinda and {Cisewski-Kehe}, Jessi and {Lissauer}, Jack J. and {Armitage}, Tayt S. and {Livesey}, Joseph R. and {Narayan}, Ritvik Sai and {Widicus Weaver}, Susanna and {Zhang}, Ke and {Bieryla}, Allyson and {Ciardi}, David R. and {Clark}, Catherine A. and {Felsmann}, Miranda and {Fernandes}, Rachel B. and {Howell}, Steve B. and {Lund}, Michael B.},
+        title = "{GJ 523b is a Massive, 170 Myr-old Mega-Earth, Likely on a Polar Orbit}",
+      journal = {arXiv e-prints},
+     keywords = {Earth and Planetary Astrophysics},
+         year = 2026,
+        month = mar,
+          eid = {arXiv:2603.24682},
+        pages = {arXiv:2603.24682},
+          doi = {10.48550/arXiv.2603.24682},
+archivePrefix = {arXiv},
+       eprint = {2603.24682},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2026arXiv260324682K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Lozovsky2026,
+       author = {{Lozovsky}, Michael},
[…]
