### 2026-04-28 aeca778
JOSS Paper bug fix
