### 2026-04-28 8f80b82
JOSS Paper bug fix
