+++ docs/paper.md
@@ -9,8 +9,8 @@ tags:
 authors:
   - name: David R. Rice
     orcid: 0000-0001-6009-8685
-    corresponding: true # (This is how to denote the corresponding author)
-    affiliation: 1
+    corresponding: true
+    affiliation: "1, 5"
   - name: Chenliang Huang
     orcid: 0000-0001-9446-6853
     affiliation: 2
@@ -18,10 +18,8 @@ authors:
     orcid: 0000-0002-0900-0192
     affiliation: 3
   - name: Mangesh Daspute
-    orcid: 
     affiliation: 4
   - name: Krishang Mittal
-    orcid: 
     affiliation: 5
   - name: Baptiste Journaux
     orcid: 0000-0002-0957-3177
@@ -44,46 +42,35 @@ affiliations:
  - name: Department of Astronomy, University of Wisconsin–Madison, 475 N. Charter Street, Madison, WI 53706, USA
    index: 5
  - name: Department of Earth and Space Science, University of Washington, Seattle, WA 98195, USA
-   index: 6 
+   index: 6
 
-date: 23 October 2025
+date: 24 April 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-Magrathea is an open-source C++ code for modeling the internal structure of differentiated planets. The initial release, @Huang:2022, introduced the base solver, a modular framework for defining equations of state (EOSs) used within phase diagrams for each differentiated layer, and outlined a series of planned extensions. Many of those updates are now implemented. Magrathea v2 is a more versatile platform that supports a wider range of compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.
-
+Magrathea is an open-source C++ code for modeling the internal structure of differentiated planets. The initial release, @Huang:2022, introduced the base solver, a modular framework for defining equations of state (EOS) within phase diagrams for each layer, and a plan for expanding the code. Many of those updates are now implemented. Magrathea v2 is a more versatile platform that supports a wider range of compositions, adds new tools for composition retrieval, and makes it easier for users to adapt the code to their own models.
 
 # Statement of need
 
-Constraining a planet’s composition is essential for understanding its formation and evolution. Observations of mass and radius alone are not sufficient, since many different interiors can yield the same bulk density. Interior structure solvers are therefore essential tools for constraining the possible compositions. With observational programs routinely measuring the densities of small to large planets, researchers require codes with models that are transparent and flexible—able to adapt to our changing understanding of planet mineralogy. 
+Constraining a planet’s composition is essential for understanding its formation and evolution. Observations of mass and radius alone are not sufficient, since many different interiors can yield the same bulk density. Interior structure solvers are therefore essential tools for constraining possible compositions. With observational programs routinely measuring the densities of small to large planets, researchers require codes with models that are transparent and flexible which can adapt as our understanding of planet mineralogy changes.
 
[…]
