+++ docs/paper.bib
@@ -294,19 +294,22 @@ archivePrefix = {arXiv},
 @ARTICLE{Steffen:2025,
        author = {{Steffen}, Jason H. and {Shakespeare}, Cody and {Royer}, Robert and {Rice}, David and {Vazan}, Allona},
         title = "{Effect of Galactic Chemical Evolution on Exoplanet Properties}",
-      journal = {arXiv e-prints},
-     keywords = {Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
+      journal = {The Astrophysical Journal Letters},
+     keywords = {Exoplanet structure, Exoplanet formation, Chemical abundances, Galaxy chemical evolution, 495, 492, 224, 580, Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
          year = 2025,
-        month = jul,
-          eid = {arXiv:2507.10942},
-        pages = {arXiv:2507.10942},
-          doi = {10.48550/arXiv.2507.10942},
+        month = oct,
+       volume = {991},
+       number = {2},
+          eid = {L33},
+        pages = {L33},
+          doi = {10.3847/2041-8213/ae0457},
 archivePrefix = {arXiv},
        eprint = {2507.10942},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250710942S},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025ApJ...991L..33S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @ARTICLE{Dorogokupets:2015,
        author = {{Dorogokupets}, P.~I. and {Dymshits}, A.~M. and {Sokolova}, T.~S. and {Danilov}, B.~S. and {Litasov}, K.~D.},
@@ -637,7 +640,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Lozovsky2026,
+@ARTICLE{Lozovsky:2026,
        author = {{Lozovsky}, Michael},
         title = "{The Uncertainty in Water Mass Fraction of Wet Planets}",
       journal = {The Astrophysical Journal},
@@ -659,7 +662,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Schulze:2026,
        author = {{Schulze}, Joseph G. and {Hinkel}, Natalie R. and {Panero}, Wendy R. and {Unterborn}, Cayman T.},
         title = "{Toward Reliable Interpretations of Small-exoplanet Compositions: Comparisons and Considerations of Equations of State and Materials Used in Common Rocky Planet Models}",
-      journal = {\psj},
+      journal = {The Planetary Science Journal},
      keywords = {Extrasolar rocky planets, Planetary interior, Exoplanet astronomy, Mineral physics, 511, 1248, 486, 2230, Earth and Planetary Astrophysics},
          year = 2026,
         month = feb,
@@ -678,7 +681,7 @@ archivePrefix = {arXiv},
 @ARTICLE{MacDonald:2022,
        author = {{MacDonald}, Mariah G. and {Feil}, Leonard and {Quinn}, Tyler and {Rice}, David},
         title = "{Confirming the 3:2 Resonance Chain of K2-138}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Exoplanet dynamics, Exoplanet migration, Exoplanet structure, 490, 2205, 495, Astrophysics - Earth and Planetary Astrophysics},
          year = 2022,
         month = apr,
