### 2026-04-28 5d58769
Updated JOSS Paper
Match new section guidelines and requirements from JOSS
