+++ docs/paper.bib
@@ -294,19 +294,22 @@ archivePrefix = {arXiv},
 @ARTICLE{Steffen:2025,
        author = {{Steffen}, Jason H. and {Shakespeare}, Cody and {Royer}, Robert and {Rice}, David and {Vazan}, Allona},
         title = "{Effect of Galactic Chemical Evolution on Exoplanet Properties}",
-      journal = {arXiv e-prints},
-     keywords = {Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
+      journal = {The Astrophysical Journal Letters},
+     keywords = {Exoplanet structure, Exoplanet formation, Chemical abundances, Galaxy chemical evolution, 495, 492, 224, 580, Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
          year = 2025,
-        month = jul,
-          eid = {arXiv:2507.10942},
-        pages = {arXiv:2507.10942},
-          doi = {10.48550/arXiv.2507.10942},
+        month = oct,
+       volume = {991},
+       number = {2},
+          eid = {L33},
+        pages = {L33},
+          doi = {10.3847/2041-8213/ae0457},
 archivePrefix = {arXiv},
        eprint = {2507.10942},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250710942S},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025ApJ...991L..33S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @ARTICLE{Dorogokupets:2015,
        author = {{Dorogokupets}, P.~I. and {Dymshits}, A.~M. and {Sokolova}, T.~S. and {Danilov}, B.~S. and {Litasov}, K.~D.},
@@ -598,6 +601,99 @@ archivePrefix = {arXiv},
         pages = {214115},
           doi = {10.1103/PhysRevB.73.214115},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2006PhRvB..73u4115L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Baumeister:2025,
+       author = {{Baumeister}, Philipp and {Miozzi}, Francesca and {Guimond}, Claire Marie and {Steinmeyer}, Marie-Luise and {Dorn}, Caroline and {Karato}, Shun-Ichiro and {Bolmont}, {\'E}meline and {Revol}, Alexandre and {Thamm}, Alexander and {Noack}, Lena},
+        title = "{Fundamentals of Interior Modelling and Challenges in the Interpretation of Observed Rocky Exoplanets}",
+      journal = {Space Science Reviews},
+     keywords = {Earth and Planetary Astrophysics, Geophysics},
+         year = 2025,
+        month = dec,
+       volume = {221},
+       number = {8},
+          eid = {123},
+        pages = {123},
+          doi = {10.1007/s11214-025-01248-5},
+archivePrefix = {arXiv},
+       eprint = {2511.10269},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025SSRv..221..123B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Kroft:2026,
[…]
