### 2026-04-28 75a2762
Update paper branch with master fixes (#36)
* Documentation: JOSS preprint & ReadTheDocs (#29)

To detail the many expansions of the code, we've included:
Preprint submitted to JOSS: https://arxiv.org/abs/2510.23733
Read The Docs: https://magrathea.readthedocs.io/en/latest/

* Update Graphite EOS

* Add files via upload

* Add temperature dependent adiabatic index

* minor change to fallback of gamma_shomate

* Update EOSlist.cpp Graphite_Lowitzer

Edit to V0 for Graphite_Lowitzer to reflect four atoms per unit cell.

* Isothermal Steam

* Isothermal steam below 100 mbar

* Update EOSlist.cpp

[…]
