+++ src/EOSlist.cpp
@@ -279,7 +279,7 @@ EOS *Water = new EOS("Water (Valencia)", Water_array, sizeof(Water_array)/2/size
 // -----------------------------------
 // Water & Super Critical, Brown 2018, Fluid Phase Equilib., 463, 18, tabulated by SeaFreeze
 // DEFAULT
-EOS *Water_Brown = new EOS("Water SC (Brown)", "./tabulated/water_Brown.txt");
+EOS *Water_Brown = new EOS("Water Liq&SC (Brown)", "./tabulated/water_Brown.txt");
 
 // -----------------------------------
 // Supercritical water. Mazevet et al. 2019, A&A 621
@@ -483,6 +483,25 @@ double vdW_H2O_array[5][2] = {{0,6}, {5,18}, {14,3}, {33, 5.537}, {34, 0.0305}};
 EOS *vdW_H2O = new EOS("H2O vdW", vdW_H2O_array, 5);
 
 // -----------------------------------
+// van der Waals, H2O, 
+//Default Hydrosphere
+double vdW_H2O_iso_array[5][2] = {{0,6}, {5,18}, {14,0}, {33, 5.537}, {34, 0.0305}};
+
+EOS *vdW_H2O_iso = new EOS("H2O vdW", vdW_H2O_iso_array, 5);
+
+//-----------------------------------
+// H2O vdW : 500–1700 K Shomate set, Chase 1998 NIST-JANAF Table
+double vdW_H2O_lo_array[][2] = {{0,6}, {5,18}, {14,3}, {33, 5.537}, {34, 0.0305}, {35, 30.09200}, {36, 6.832514}, {37, 6.793435}, {38, -2.534480}, {39, 0.082139}};
+
+EOS *vdW_H2O_lo = new EOS("H2O vdW",vdW_H2O_lo_array,sizeof(vdW_H2O_lo_array)/2/sizeof(vdW_H2O_lo_array[0][0]));
+
+//-----------------------------------
+// H2O vdW : 1700–6000 K Shomate set, Chase 1998 NIST-JANAF Table
+double vdW_H2O_hi_array[][2] = {{0,6}, {5,18}, {14,3}, {33, 5.537}, {34, 0.0305}, {35, 41.96426}, {36, 8.622053}, {37, -1.499780}, {38, 0.098119}, {39, -11.15764}};
+
+EOS *vdW_H2O_hi = new EOS("H2O vdW",vdW_H2O_hi_array,sizeof(vdW_H2O_lo_array)/2/sizeof(vdW_H2O_lo_array[0][0]));
+
+// -----------------------------------
 // van der Waals, CH4, 
 double vdW_CH4_array[5][2] = {{0,6}, {5,16}, {14,3}, {33, 2.303}, {34, 0.0431}};
 
@@ -536,6 +555,7 @@ EOS *Graph = new EOS("Graph",Graph_array,sizeof(Graph_array)/2/sizeof(Graph_arra
 double Diamond_array[][2] = {{0,2},{1,3.43467},{2,432.4},{3,3.793},{5,mC},{7,1887.8},{8,0.5836},{9,1},{10,0.499},{14,1},{15,6}};
 
 EOS *Diam = new EOS("Diamond",Diamond_array,sizeof(Diamond_array)/2/sizeof(Diamond_array[0][0]));
+
 // -----------------------------------
 // BC8 (Vinet), Benedict et al. 2018, Phys  Rev B
 // Uses the 1 terms from the double Debye model found in Benedict et al (2018)
@@ -544,6 +564,13 @@ double BC8_array[][2] = {{0,2},{1,3.75902},{2,221.2},{3,4.697},{5,mC},{7,2800.6}
 
[…]

[…]
