### 2026-04-28 #9300 @DavidRRice
@warrickball the JOSS manuscript has been updated to meet the new guidelines.
