### 2026-05-03 #9300 @warrickball
Thanks for the update. I just want to confirm that this is certainly in scope and can return to the queue for an editor, which is still long. I'm not confident enough to predict how much longer you might be waiting since it now depends increasingly on the random nature of how long other ongoing reviews take to finish up. I can only apologise once again and hope that it's not too long now.
