### 2026-07-11 7e71d58
Update affiliation for David R. Rice in README
