### 2026-07-31 f07b56d
Documentation Updates
-New Logo! :)
-Community Guidelines in CONTRIBUTING.md
-Tutorial linked and recreated in RTD
-JOSS paper edit
-Quick Start instructions improvement
