+++ docs/rtd/getting-started.md
@@ -1,27 +1,27 @@
 # Installation and First Run
 
-MAGRATHEA is an open-source **1-D planetary interior structure solver** written in C++.  
+Magrathea is an open-source **1-D planetary interior structure solver** written in C++.  
 It is designed for astronomers, planetary scientists, and students who want to model the internal structure of differentiated planets.
 
----
+\---
 
 ## 🪐 Overview
 
-MAGRATHEA integrates the hydrostatic equilibrium equations to compute the radius, pressure, temperature, density, and phase structure of a planet consisting of up to **four layers**:
+Magrathea integrates the hydrostatic equilibrium equations to compute the radius, pressure, temperature, density, and phase structure of a planet consisting of up to **four layers**:
 
-- **Core** – iron or alloy  
-- **Mantle** – silicate  
-- **Hydrosphere** – water/ice  
-- **Atmosphere** – ideal gas  
+* **Core** – iron or alloy
+* **Mantle** – silicate
+* **Hydrosphere** – water/ice
+* **Atmosphere** – ideal gas
 
 Each layer uses a modular **phase diagram** and **equation of state (EOS)** that can be modified or replaced.  
 The solver outputs the full radial structure and layer boundaries, allowing you to explore how composition or temperature affect planetary size and density.
 
----
+\---
 
 ## ⚙️ 1. Prerequisites
 
-MAGRATHEA requires the **GNU Scientific Library (GSL ≥ 2.0)**.
+Magrathea requires the **GNU Scientific Library (GSL ≥ 2.0)**.
 
 ### Linux / macOS
 
@@ -46,7 +46,7 @@ sudo make install
 Use **Windows Subsystem for Linux (WSL)** and follow the same Linux instructions inside the terminal.  
 (Direct compilation on native Windows is not supported because of path issues.)
 
----
+\---
 
 ## 📦 2. Getting the Code
 
@@ -71,56 +71,57 @@ gsl-config --cflags
 gsl-config --libs
 ```
 
----
+\---
 
-## 🧱 3. Building MAGRATHEA
+## 🧱 3. Building Magrathea
 
-Run:
-
-```bash
-make -B        # first build
-make           # subsequent builds
-```
-
-This creates an executable named **`planet`** in the project directory.
-
----
-
-## 🚀 4. Running a Test Planet
-
-The simplest run mode is `mode0.cfg`, which solves for a planet’s radius and interior given its layer masses and temperatures.
-
-1. Open `run/mode0.cfg` in a text editor.  
-2. Adjust these key parameters:
-
-| Lines | Parameter | Description |
-|-------|------------|-------------|
-| 12–14 | `mass_of_core`, `mass_of_mantle`, `mass_of_hydro`, `mass_of_atm` | Layer masses (Earth masses) |
-| 16–20 | `surface_temp`, `temp_jump_1–3` | Surface temperature and discontinuities between layers |
-| 21 | `output_file` | Output file name and path |
-
-**Example**
-
-```cfg
-mass_of_core = 0.33
-mass_of_mantle = 0.67
-mass_of_hydro = 0.0
-mass_of_atm = 0.0
-
-surface_temp = 300
-temp_jump_1 = 0
-temp_jump_2 = 0
-temp_jump_3 = 0
-
-output_file = result/earthlike.txt
-```
-
-3. Compile and run:
+From the top-level Magrathea directory, run:
 
 ```bash
-make
+make -B
+```
+
+This creates an executable named **`planet`** in the project directory. After the first compilation, use `make` when C++ source files have changed.
+
[…]
