+++ README.md
@@ -2,7 +2,7 @@
 
 [<img src="plot/Magrathea_Logo.png" width="350"/>](plot/Magrathea_Logo.png)
 
-**Excerpt from The Hitchhiker's Guide to the Galaxy, Page 634784, Section 5a, Entry: MAGRATHEA**
+**Excerpt from The Hitchhiker's Guide to the Galaxy, Page 634784, Section 5a, Entry: Magrathea**
 
 Planet interior structure code for astronomers, planetary scientists, mice, and more. 
 
@@ -16,7 +16,7 @@ The code returns the pressure, temperature, density, phase, and radius at steps 
 The code supports 4 layers: core, mantle, hydrosphere, and atmosphere. Each layer has a phase diagram with equations of state (EoS) chosen for each phase.
 The code was developed by [Chenliang Huang](https://huang-cl.github.io/), [David R. Rice](https://davidrrice.github.io/), and [Jason H. Steffen](https://www.jasonhsteffen.com/) at the Univerisity of Nevada, Las Vegas starting in 2017. For further documentation visit our Read the Docs and our publications above. A tutorial and practice projects/problems are found in [docs/Tutorial_Practice_Problems.pdf](docs/Tutorial_Practice_Problems.pdf)
 
-We encourage the community to contribute to and use MAGRATHEA for their interior modeling needs.
+We encourage the community to contribute to and use Magrathea for their interior modeling needs.
 
  <p align="center">
 <img width = "600" src="plot/funplanets.png"/>
@@ -28,7 +28,7 @@ We welcome bug reports, documentation improvements, equations of state, phase di
 
 * See [CONTRIBUTING.md](CONTRIBUTING.md) for contribution instructions and pull-request guidance.
 * Use [GitHub Issues](https://github.com/Huang-CL/Magrathea/issues) to report problems, request features, or ask for help.
-* See [CITATION.md](CITATION.md) for citation instructions and [citations.md](citations.md) for a list of published works using MAGRATHEA.
+* See [CITATION.md](CITATION.md) for citation instructions and [citations.md](citations.md) for a list of published works using Magrathea.
 
 ## Prerequisite ##
 
@@ -50,24 +50,89 @@ When running many simulations with bulk input or composition finder modes, OpenM
 
 Installation:
 
-1. Open a terminal. Navigate to where you wish to install the directory.
-2. Clone the repository: `git clone https://github.com/Huang-CL/Magrathea.git`.
-3. If the gsl library is not installed globally (under /usr/local/ or equivalent), edit `Makefile` of MAGRATHEA to include the actual path toward the gsl headers (e.g. `~/include` or `~/gsl/include`) and gsl library (e.g. `~/lib` or `~/gsl/lib`) following `-I` in `CFLAGS` and `-L` in `LDFLAGS`. The path of headers and libraries can be found using `gsl-config --cflags` and `gsl-config --libs`.  If the `gsl-config` command is not found, `gsl` may not have been installed properly.  _Note: The path after `-I` should end with `/include`, not `/include/gsl`._
-4. Run `make -B` inside the code's directory. _After the first compilation use `make` to compile the code._
+1. Open a terminal and navigate to where you wish to install Magrathea.
 
-Running a planet:
+2. Clone the repository and enter its top-level directory:
 
-5. Open `run/mode0.cfg` in a text editor.
-6. Line 12-14: set mass of the core, mantle, hydrosphere, and atmosphere in Earth-masses.
-7. Line 16-20: set surface temperature and any temperature discontinuities.
-8. Line 21: set output file name and path.
-9. In terminal, compile changed file with `make`.
-10. Run MAGRATHEA with `./planet run/mode0.cfg`.
[…]
