### 2026-07-31 cda1aaf
Update paper.md
Point Figure 1 caption to PhaseDiagrams repository for reproducibility
