### 2026-07-31 87345eb
Implement Automated Testing
Implemented command "./planet --test" which tests the EOS inverter on an analytical Vinet EOS and runs a constant density planet checking the solved radius verses the analytical radius.

Describe test procedure in README,  RTD, and the JOSS paper
