### 2026-07-31 0a7a71d
Updates for JOSS Revision, Automated Tests (#37)
* Documentation Updates

-New Logo! :)
-Community Guidelines in CONTRIBUTING.md
-Tutorial linked and recreated in RTD
-JOSS paper edit
-Quick Start instructions improvement

* Update paper.md

Point Figure 1 caption in JOSS publication to PhaseDiagrams repository for reproducibility

* Implement Automated Testing

Implemented command "./planet --test" which tests the EOS inverter on an analytical Vinet EOS and runs a constant density planet checking the solved radius verses the analytical radius.

Describe test procedure in README,  RTD, and the
[…]
