+++ citations.md
@@ -3,15 +3,17 @@
 [Citation Instructions](CITATION.md)
 
 
-## Journal Articles
+## Journal Articles using Magrathea
 
-10. Steffen J. H., Shakespeare C., Royer R., Rice D., Vazan A, 2025, Effect of Galactic Chemical Evolution on Exoplanet Properties, ApJ 991. doi:[10.3847/2041-8213/ae0457](https://doi.org/10.3847/2041-8213/ae0457). ([ADS](https://ui.adsabs.harvard.edu/abs/2025ApJ...991L..33S))
-9. Taylor J. and 20 colleagues, 2025, JWST NIRISS Transmission Spectroscopy of the Super-Earth GJ 357b, a Favourable Target for Atmospheric Retention, MNRAS, doi:[10.1093/mnras/staf894](https://doi.org/10.1093/mnras/staf894). ([ADS](https://ui.adsabs.harvard.edu/abs/2025MNRAS.tmp..888T))
-8. Rice D. R., Huang C., Steffen J. H., Vazan A., 2025, Uncertainties in the Inference of Internal Structure: The Case of TRAPPIST-1 f, ApJ, 986, doi:[10.3847/1538-4357/add34b](https://doi.org/10.3847/1538-4357/add34b). ([ADS](https://ui.adsabs.harvard.edu/abs/2025ApJ...986....2R))
-7. Daspute M., Wandel A., Kopparapu R. K., Perdelwitz V., Teklu J. T., Tal-Or L., 2025, Potential Interior Structures and Habitability of Super-Earth Exoplanets LHS 1140 b, K2-18 b, TOI-1452 b, and TOI-1468 c, ApJ, 979. doi:[10.3847/1538-4357/ad9ba9](https://doi.org/10.3847/1538-4357/ad9ba9). ([ADS](https://ui.adsabs.harvard.edu/abs/2025ApJ...979..158D))
-6. Desai A. and 39 colleagues, 2024, The TESS-Keck Survey. XVIII. A sub-Neptune and spurious long-period signal in the TOI-1751 system, ApJ, 167, doi:[10.3847/1538-3881/ad29ee](https://doi.org/10.3847/1538-3881/ad29ee). ([ADS](https://ui.adsabs.harvard.edu/abs/2024AJ....167..194D))
-5. Dou J., Carter P. J., Leinhardt, Z. M., 2024, Formation of super-Mercuries via giant impacts, MNRAS, 529, 2577 doi:[10.1093/mnras/stae644](https://doi.org/10.1093/mnras/stae644). ([ADS](https://ui.adsabs.harvard.edu/abs/2024MNRAS.tmp..686D))
-4. Childs A. C., Shakespeare C., Rice D. R., Yang C. C., Steffen, J. H., 2023, Composition constraints of the TRAPPIST-1 planets from their formation, MNRAS, 524, 3749–3768. doi:[10.1093/mnras/stad2110](https://doi.org/10.1093/mnras/stad2110). ([ADS](https://ui.adsabs.harvard.edu/abs/2023MNRAS.524.3749C))
-3. Huang C., Rice D. R., Steffen J. H., 2022, MAGRATHEA: an open-source spherical symmetric planet interior structure code, MNRAS, 513, 5256–5269. doi:[10.1093/mnras/stac1133](https://doi.org/10.1093/mnras/stac1133). ([ADS](https://ui.adsabs.harvard.edu/abs/2022arXiv220103094H))
[…]
