### 2026-07-31 77d62e6
Add references to bib
Add Wang+19 and Attia+26
