### 2026-07-31 56830e8
Add new articles to bibliography
