+++ docs/paper.bib
@@ -697,3 +697,37 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Wang:2019,
+       author = {{Wang}, H.~S. and {Liu}, F. and {Ireland}, T.~R. and {Brasser}, R. and {Yong}, D. and {Lineweaver}, C.~H.},
+        title = "{Enhanced constraints on the interior composition and structure of terrestrial exoplanets}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {planets and satellites: composition, planets and satellites: interiors, planets and satellites: terrestrial planets, stars: abundances, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2019,
+        month = jan,
+       volume = {482},
+       number = {2},
+        pages = {2222-2233},
+          doi = {10.1093/mnras/sty2749},
+archivePrefix = {arXiv},
+       eprint = {1810.04615},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.2222W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Attia:2026,
+       author = {{Attia}, Mara and {Lichtenberg}, Tim and {Jungov{\'a}}, Ema and {Sastre}, Mariana},
+        title = "{PALEOS: Multiphase equations of state and mass-radius relations for exoplanet interiors}",
+      journal = {arXiv e-prints},
+     keywords = {Earth and Planetary Astrophysics},
+         year = 2026,
+        month = may,
+          eid = {arXiv:2605.03741},
+        pages = {arXiv:2605.03741},
+          doi = {10.48550/arXiv.2605.03741},
+archivePrefix = {arXiv},
+       eprint = {2605.03741},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2026arXiv260503741A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
