### 2026-08-18 #10692 @DavidRRice
Hi @TobiBu, checking in if the editing team needs anything more on your end. The new pdf generated. Is it all good for a second round of review?
