### 2026-09-03 #10692 @TobiBu
Hi @DavidRRice,

I was just checking the License and your repo runs under MIT while the zenodo archive is Creative Commons.
Can you changed that?
