### 2026-09-03 #10692 @DavidRRice
@TobiBu Corrected. Thank you for noticing. Zenodo matches MIT license now: https://zenodo.org/records/22182485
