### 2026-09-03 #10692 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
